pyproject.toml

[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord.py[voice]>=2.3.0",
    "yt-dlp>=2023.1.6",
    "spotipy>=2.22.1",
    "PyNaCl>=1.5.0",
]
